Update README.md
Thank you for such a great library!
In this PR I'm fixing a couple small things I noticed while reading the Readme.

I'm planning to use this library on my personal website, and one thing I noticed is that the current NPM release does not seem to have TS types. And I see that they have been in the repo for quite awhile now. Is there something missing there? What's blocking the release? I'd be happy to help.